one-month-finish

+++ app/controllers/requests_controller.rb
@@ -1,7 +1,6 @@
 class RequestsController < ApplicationController
     
 	def create
-			debugger
 		@user = User.find_by(name: params[:request][:superior])
 		@request = @user.requests.new(request_date: params[:request_date], category: params[:category], applicant: params[:applicant], finish_time: params[:request]["finish_time(3i)"], note: params[:note], change_date: params[:change_date])
 		@request.save
@@ -13,35 +12,47 @@ class RequestsController < ApplicationController
 			@status.update(status3: 1)
 		end
 		redirect_to user_path
-		
 	end
 	
-	def destroy
 	
+	
+	def destroy
 		@statuses = params[:status]
 		@user = User.find(params[:id])
 		@requests = @user.requests.where(category: 1).order(:applicant)
 		s = 0
+		m = 0
+		pre_applicant = @requests.first.applicant
+		
 		@requests.each do |r|
-			if @statuses[s] == "2"
-				@applied_user = User.find(r.applicant).attendances.find_by(worked_on: r.request_date)
-				@applied_user.update(status1: 2)
-				r.destroy
-			elsif @statuses[s] == "3"
-				@applied_user = User.find(r.applicant).attendances.find_by(worked_on: r.request_date)
-				@applied_user.update(status1: 3)
-				r.destroy
+			if pre_applicant != r.applicant
+				m = 0
+			end
+			if params["check_box#{r.applicant}"].present?
+				if params["check_box#{r.applicant}"]["#{m}"] == "1"
+					if @statuses[s] == "2"
+						@applied_user = User.find(r.applicant).attendances.find_by(worked_on: r.request_date)
+						@applied_user.update(status1: 2)
+						r.destroy
+					elsif @statuses[s] == "3"
+						@applied_user = User.find(r.applicant).attendances.find_by(worked_on: r.request_date)
+						@applied_user.update(status1: 3)
+						r.destroy
+					end
+				end
 			end
 			s += 1
+			m += 1
+			pre_applicant = r.applicant
 		end
-		redirect_to user_path
 		
+		redirect_to user_path
 	end
 	
 	private
 	
-	def params_request
-		params.require(:request).permit(:request_month, :applicant, :category)
-	end
-	
+		def params_request
+			params.require(:request).permit(:request_month, :applicant, :category)
+		end
+		
 end
